# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Deliver Lubumbashi - Food Delivery App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List all stores
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a store
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{id}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{id}/menu:
    get:
      operationId: getStoreMenu
      tags: [menu]
      summary: Get menu for a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Add a menu item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemRequest"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    Store:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        address:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        deliveryTimeMin:
          type: integer
        deliveryFee:
          type: number
        isOpen:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ownerId
        - name
        - category
        - address
        - phone
        - rating
        - deliveryTimeMin
        - deliveryFee
        - isOpen
        - createdAt
    CreateStoreRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        address:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
        deliveryTimeMin:
          type: integer
        deliveryFee:
          type: number
      required:
        - name
        - category
        - address
        - phone
        - deliveryTimeMin
        - deliveryFee
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        storeId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        isAvailable:
          type: boolean
      required:
        - id
        - storeId
        - name
        - price
        - category
        - isAvailable
    CreateMenuItemRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
      required:
        - name
        - price
        - category
...